+++ src/components/base/index.js
@@ -11,3 +11,4 @@ export { default as Modal } from './Modal/Modal'
 export { default as RadioButton } from './RadioButton/RadioButton'
 export { default as CheckBox } from './CheckBox/CheckBox'
 export { default as InnerDeleteButton } from './InnerDeleteButton/InnerDeleteButton'
+export { default as Gnb } from './Gnb/Gnb'
