+++ src/pages/ProductEnrollAdmin/style.jsx
@@ -19,33 +19,6 @@ export const Logo = styled.h1`
   z-index: 2;
 `
 
-export const Gnb = styled.ul`
-  width: 15vw;
-  min-width: 180px;
-  border-right: 1px solid ${({ theme }) => theme.color.border};
-  position: fixed;
-  top: 50px;
-  left: 0;
-`
-
-export const Menu = styled.li`
-  width: 100%;
-  padding: ${({ theme }) => theme.table.padding};
-  font: bold 16px/1 'arial';
-  color: ${({ theme }) => theme.color.primary};
-  border-bottom: 1px solid ${({ theme }) => theme.color.border};
-  cursor: pointer;
-`
-
-export const InnerMenu = styled(Menu)`
-  border-bottom: none;
-  background: ${({ theme }) => theme.color.menuBackground};
-`
-export const ActiveMenu = styled(InnerMenu)`
-  color: ${({ theme }) => theme.color.purpleDarker};
-  background: ${({ theme }) => theme.color.selectedBackground};
-`
-
 export const FormWrapper = styled.div`
   width: 100%;
   position: relative;
